fix conda build
