+++ assets/stable-content/.BASE_VERSION(STABLE)
