+++ assets/stable-content/_Dockerfile-base-STABLE
