+++ assets/stable-content/_Dockerfile-baseSTABLE
