+++ assets/stable-content/_Dockerfile-base_BIOSIMULATOR_PROCESSES(original)
