--- worker/pyproject.toml
@@ -1,16 +1,0 @@
-[project]
-name = "worker"
-version = "0.1.0"
-description = "Worker container for Simulation Verification Service Microservice"
-authors = [{name = "Alex Patrie", email = "alexanderpatrie@gmail.com"}]
-readme = "README.md"
-requires-python = ">=3.10"
-
-
-[tool.setuptools.packages.find]
-include = ["service"]
-
-
-[build-system]
-requires = ["setuptools", "wheel"]
-build-backend = "setuptools.build_meta"
