--- worker/service/__init__.py
@@ -1,6 +1,0 @@
-
-
-
-# APP_PROCESS_REGISTRY.process_registry.register('time-course-output-generator', TimeCourseOutputGenerator)
-# APP_PROCESS_REGISTRY.process_registry.register('output-generator', OutputGenerator)
-
