+++ worker/simularium_utils.py
